requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain
langchain-openai
openai==2.18.0
gspread==5.12.0
oauth2client==4.1.3
pandas==2.1.3
streamlit
python-dotenv
python-dateutil
